+++ UniTAS/Plugin/Services/UnitySafeWrappers/Wrappers/ITimeWrapper.cs
@@ -6,7 +6,7 @@ public interface ITimeWrapper
     /// Sets the game's frame time between updates
     /// If IntFPSOnly is true, this will be rounded to the nearest integer
     /// </summary>
-    float CaptureFrameTime { get; set; }
+    double CaptureFrameTime { get; set; }
 
     /// <summary>
     /// Returns if frame time can only be FPS values of an integer
